chore: optimize plugin code directory structure
