+++ packages/plugins/robot/index.ts
@@ -14,7 +14,7 @@ import entry from './src/Main.vue'
 import metaData from './meta'
 import './src/styles/vars.less'
 import '@opentiny/tiny-robot/dist/style.css'
-import { RobotService } from './src/composables/index'
+import { RobotService } from './src/metas'
 
 export default {
   ...metaData,
