--- packages/plugins/robot/src/components/LoadingRenderer.vue
@@ -1,3 +1,0 @@
-<template>
-  <img src="../../assets/loading.webp" alt="loading" style="width: 24px; height: 24px" />
-</template>
