+++ packages/plugins/robot/src/components/chat/FooterButton.vue
