+++ packages/plugins/robot/src/components/footer-extension/RobotTypeSelect.vue
