+++ packages/plugins/robot/src/components/icons/mcp-icon.vue
