+++ packages/plugins/robot/src/components/icons/page-icon.vue
