+++ packages/plugins/robot/src/components/icons/study-icon.vue
