+++ packages/plugins/robot/src/components/renderers/ImgRenderer.vue
