+++ packages/plugins/robot/src/components/renderers/LoadingRenderer.vue
@@ -1,0 +1,3 @@
+<template>
+  <img src="../../../assets/loading.webp" alt="loading" style="width: 24px; height: 24px" />
+</template>
