+++ packages/plugins/robot/src/constants/index.ts
@@ -1,0 +1,1 @@
+export * from './model-config'
