+++ packages/plugins/robot/src/metas/index.ts
@@ -1,10 +1,13 @@
 import { HOOK_NAME } from '@opentiny/tiny-engine-meta-register'
-import useRobot from './useRobot'
+import useModelConfig from '../composables/useConfig'
 
 export const RobotService = {
   id: 'engine.service.robot',
   type: 'MetaService',
-  apis: useRobot(),
+  apis: {
+    robotSettingState: useModelConfig().robotSettingState,
+    getAIModelOptions: useModelConfig().getAIModelOptions
+  },
   composable: {
     name: HOOK_NAME.useRobot
   }
