+++ packages/plugins/robot/src/prompts/data/components.json
