+++ packages/plugins/robot/src/prompts/data/examples.json
