+++ packages/plugins/robot/src/prompts/templates/chat-prompt.md
