+++ packages/plugins/robot/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './mcp.types'
+export * from './chat.types'
